Create-JWT-service: move JwtEntityFactory to factory package
